Moving all necessary files to utils
