+++ app/traffic_cloud_utils/plotting/car.png
