+++ app/traffic_cloud_utils/plotting/pedestrian.png
